requirements.txt
fastapi
uvicorn
python-dotenv
httpx
torch
transformers
scikit-learn
Pillow
requests
beautifulsoup4
